+++ core/telemetry/__init__.py
@@ -2,20 +2,22 @@
 
 from .client import (
     ALLOWED_EVENT_NAMES,
-    MODULE_EVENT_NAMES,
     TelemetryClient,
     clear_telemetry_queue,
+    emit_startup_telemetry,
     emit_telemetry,
     flush_telemetry,
     start_telemetry_flush,
+    telemetry_status,
 )
 
 __all__ = [
     "ALLOWED_EVENT_NAMES",
-    "MODULE_EVENT_NAMES",
     "TelemetryClient",
     "clear_telemetry_queue",
+    "emit_startup_telemetry",
     "emit_telemetry",
     "flush_telemetry",
     "start_telemetry_flush",
+    "telemetry_status",
 ]